+++ decorators/index.js
@@ -1,1 +1,2 @@
 export { default as controllersWrapper } from "./controllersWrapper.js";
+export { default as validateBody } from "./validateBody.js";
