+++ models/contacts.json
@@ -64,5 +64,23 @@
     "name": "Samson",
     "email": "mail@mail.com",
     "phone": "123-123-129"
+  },
+  {
+    "id": "qMGD0H8eIz1_S0RnOt6ju",
+    "name": "Anna",
+    "email": "mail@mail.com",
+    "phone": "123-444-222"
+  },
+  {
+    "id": "EU6QLG6aI7bn8coOoHJsM",
+    "name": "Anna",
+    "email": "mail@mail.com",
+    "phone": "123-444-222"
+  },
+  {
+    "id": "zptEMSZbyB6OqAALo-MxC",
+    "name": "Victor",
+    "email": "mail@mail.com",
+    "phone": "123-444-222"
   }
 ]
