+++ routes/api/contacts-router.js
@@ -1,6 +1,10 @@
 import express from "express";
 import contactsControllers from "../../controllers/contacts-controllers.js";
 import { isEmptyBody } from "../../middlewares/index.js";
+import { contactAddSchema } from "../../schemas/index.js";
+import { validateBody } from "../../decorators/index.js";
+
+const contactAddValidate = validateBody(contactAddSchema);
 
 const router = express.Router();
 
@@ -8,10 +12,15 @@ router.get("/", contactsControllers.getAll);
 
 router.get("/:contactId", contactsControllers.getById);
 
-router.post("/", isEmptyBody, contactsControllers.addNew);
+router.post("/", isEmptyBody, contactAddValidate, contactsControllers.addNew);
 
 router.delete("/:contactId", contactsControllers.removeById);
 
-router.put("/:contactId", isEmptyBody, contactsControllers.updateById);
+router.put(
+  "/:contactId",
+  isEmptyBody,
+  contactAddValidate,
+  contactsControllers.updateById
+);
 
 export default router;
